Migrate some tests to new subcrate
