+++ gen-models/src/block_group.rs
@@ -1239,16 +1239,12 @@ impl Query for BlockGroup {
 
 #[cfg(test)]
 mod tests {
-    use core::ops::Range;
-    use std::collections::HashSet;
-
     use capnp::message::TypedBuilder;
     use chrono::Utc;
-    use gen_core::{NO_CHROMOSOME_INDEX, region::RegionResolutionError};
+    use gen_core::region::RegionResolutionError;
 
     use super::*;
     use crate::{
-        annotations::Annotation as ModelAnnotation,
         collection::Collection,
         node::Node,
         region::{ResolvedGenRegion, ResolvedRegionKind},
@@ -1450,186 +1446,6 @@ mod tests {
         assert_eq!(
             BlockGroupEdge::edges_for_block_group(conn, &bg1.id, None),
             BlockGroupEdge::edges_for_block_group(conn, &bg2, None)
-        );
-    }
-
-    #[test]
-    fn test_blockgroup_copies_immediate_parent_block_groups() {
-        let conn = &get_connection(None).unwrap();
-        Collection::create(conn, "test").unwrap();
-        Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "parent_a",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "parent_b",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "child",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-
-        let parent_a_bg = create_bg(conn, "test", "parent_a", "chr1");
-        let parent_b_bg = create_bg(conn, "test", "parent_b", "chr1");
-
-        let seq_a = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("AAAA")
-            .save(conn)
-            .unwrap();
-        let seq_b = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("CCCC")
-            .save(conn)
-            .unwrap();
-        let node_a =
-            Node::create(conn, &seq_a.hash, &HashId::convert_str("merge-parent-a")).unwrap();
-        let node_b =
-            Node::create(conn, &seq_b.hash, &HashId::convert_str("merge-parent-b")).unwrap();
-
-        let parent_a_edges = [
-            Edge::create(
-                conn,
-                PATH_START_NODE_ID,
-                0,
-                Strand::Forward,
-                node_a,
-                0,
-                Strand::Forward,
-            )
-            .unwrap(),
-            Edge::create(
-                conn,
-                node_a,
-                4,
-                Strand::Forward,
-                PATH_END_NODE_ID,
-                0,
-                Strand::Forward,
-            )
-            .unwrap(),
-        ];
-        let parent_b_edges = [
-            Edge::create(
-                conn,
-                PATH_START_NODE_ID,
-                0,
-                Strand::Forward,
-                node_b,
-                0,
-                Strand::Forward,
-            )
-            .unwrap(),
-            Edge::create(
-                conn,
-                node_b,
-                4,
-                Strand::Forward,
-                PATH_END_NODE_ID,
-                0,
-                Strand::Forward,
-            )
-            .unwrap(),
-        ];
-
-        BlockGroupEdge::bulk_create(
-            conn,
-            &parent_a_edges
-                .iter()
-                .map(|edge| BlockGroupEdgeData {
-                    block_group_id: parent_a_bg.id,
-                    edge_id: edge.id,
-                    chromosome_index: 0,
-                    phased: 0,
-                })
-                .collect::<Vec<_>>(),
-        );
-        BlockGroupEdge::bulk_create(
-            conn,
-            &parent_b_edges
-                .iter()
-                .map(|edge| BlockGroupEdgeData {
-                    block_group_id: parent_b_bg.id,
-                    edge_id: edge.id,
-                    chromosome_index: 0,
-                    phased: 0,
-                })
-                .collect::<Vec<_>>(),
-        );
-
-        let child_block_groups = BlockGroup::get_or_create_sample_block_groups(
-            conn,
-            "test",
-            "child",
-            "chr1",
-            vec!["parent_a".to_string(), "parent_b".to_string()],
-        )
-        .unwrap();
-        assert_eq!(child_block_groups.len(), 2);
-
-        let child_by_parent = child_block_groups
-            .iter()
-            .map(|block_group| (block_group.parent_block_group_id.unwrap(), block_group))
-            .collect::<HashMap<_, _>>();
-
-        let child_a = child_by_parent.get(&parent_a_bg.id).unwrap();
-        let child_b = child_by_parent.get(&parent_b_bg.id).unwrap();
-
-        let child_a_edges = BlockGroupEdge::query(
-            conn,
-            "select * from block_group_edges where block_group_id = ?1",
-            params![child_a.id],
-        );
-        let child_b_edges = BlockGroupEdge::query(
-            conn,
-            "select * from block_group_edges where block_group_id = ?1",
-            params![child_b.id],
-        );
-        assert_eq!(
-            child_a_edges
-                .iter()
-                .map(|edge| edge.edge_id)
-                .collect::<HashSet<_>>(),
-            parent_a_edges
-                .iter()
-                .map(|edge| edge.id)
-                .collect::<HashSet<_>>()
-        );
-        assert_eq!(
-            child_b_edges
-                .iter()
-                .map(|edge| edge.edge_id)
-                .collect::<HashSet<_>>(),
-            parent_b_edges
-                .iter()
-                .map(|edge| edge.id)
-                .collect::<HashSet<_>>()
-        );
-
-        assert_eq!(
-            BlockGroup::get_all_sequences(conn, &child_a.id, false).unwrap(),
-            HashSet::from_iter(vec!["AAAA".to_string()])
-        );
-        assert_eq!(
-            BlockGroup::get_all_sequences(conn, &child_b.id, false).unwrap(),
-            HashSet::from_iter(vec!["CCCC".to_string()])
-        );
-        assert_eq!(
-            Sample::get_all_sequences(conn, "test", "child", false, None).unwrap(),
-            HashSet::from_iter(vec!["AAAA".to_string(), "CCCC".to_string()])
         );
     }
 
@@ -2175,899 +1991,6 @@ mod tests {
     }
 
     #[test]
-    fn insert_accession_change_get_all() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let mut path_cache = PathCache::new(&conn);
-        let accession =
-            BlockGroup::add_accession(&conn, &path, "test-accession", 10, 30, &mut path_cache)
-                .unwrap();
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id = Node::create(
-            &conn,
-            &insert_sequence.hash,
-            &HashId::convert_str("acc-insert-node"),
-        )
-        .unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 5,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_accession(&conn, &accession, 5, 15).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-
-        BlockGroup::insert_change(&conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTNNNNCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_annotation_change_get_all() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let mut path_cache = PathCache::new(&conn);
-        let accession =
-            BlockGroup::add_accession(&conn, &path, "test-accession", 10, 30, &mut path_cache)
-                .unwrap();
-        let annotation =
-            ModelAnnotation::get_or_create(&conn, "gene-1", "track-1", &accession.id, None)
-                .unwrap();
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id = Node::create(
-            &conn,
-            &deletion_sequence.hash,
-            &HashId::convert_str("annotation-delete-node"),
-        )
-        .unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 5,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let annotation_accession =
-            Accession::get_by_id(&conn, &annotation.accession_id, None).unwrap();
-        let region =
-            ResolvedGenRegion::from_annotation(&conn, &annotation, &annotation_accession, 5, 15)
-                .unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-
-        BlockGroup::insert_change(&conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_and_deletion_get_all() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 7, 15).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id =
-            Node::create(&conn, &deletion_sequence.hash, &HashId::convert_str("2")).unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 19,
-            path_end: 31,
-            strand: Strand::Forward,
-        };
-
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 19, 31).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn simple_insert_get_all() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 7, 15).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_on_block_boundary_middle() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 15,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 15, 15).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTNNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_within_block() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 12,
-            path_end: 17,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 12, 17).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTNNNNTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_on_block_boundary_start() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 10,
-            path_end: 10,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 10, 10).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAANNNNTTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_on_block_boundary_end() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 9,
-            path_end: 9,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 9, 9).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAANNNNATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_across_entire_block_boundary() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 10,
-            path_end: 20,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 10, 20).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAANNNNCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_across_two_blocks() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 15,
-            path_end: 25,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 15, 25).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTNNNNCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_spanning_blocks() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 5,
-            path_end: 35,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 5, 35).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAANNNNGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn simple_deletion() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id =
-            Node::create(&conn, &deletion_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 19,
-            path_end: 31,
-            strand: Strand::Forward,
-        };
-
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 19, 31).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-
-        // take out an entire block
-        BlockGroup::insert_change(&conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn doesnt_apply_same_insert_twice() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 7, 15).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_at_beginning_of_path() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 0,
-            path_end: 0,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 0, 0).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "NNNNAAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn homozygous_insert_at_beginning_of_path() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 0,
-            path_end: 0,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 0, 0).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 0,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "NNNNAAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_at_end_of_path() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 40,
-            path_end: 40,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 40, 40).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGGNNNN".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_at_one_bp_into_block() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 10,
-            path_end: 11,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 10, 11).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAANNNNTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn insert_at_one_bp_from_end_of_block() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(&conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(&conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 19,
-            path_end: 20,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 19, 20).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTNNNNCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn delete_at_beginning_of_path() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id =
-            Node::create(&conn, &deletion_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 0,
-            path_end: 1,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 0, 1).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn delete_at_end_of_path() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id =
-            Node::create(&conn, &deletion_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 35,
-            path_end: 40,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 35, 40).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
     fn error_on_out_of_bounds_change() {
         let conn = get_connection(None).unwrap();
         let (block_group_id, path) = setup_block_group(&conn);
@@ -3111,88 +2034,6 @@ mod tests {
         assert!(matches!(res, Err(BlockGroupError::ChangeOutOfBounds(_))));
         let res = BlockGroup::insert_change(&conn, &before_start_change);
         assert!(matches!(res, Err(BlockGroupError::ChangeOutOfBounds(_))));
-    }
-
-    #[test]
-    fn deletion_starting_at_block_boundary() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id =
-            Node::create(&conn, &deletion_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 10,
-            path_end: 12,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 10, 12).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-    }
-
-    #[test]
-    fn deletion_ending_at_block_boundary() {
-        let conn = get_connection(None).unwrap();
-        let (block_group_id, path) = setup_block_group(&conn);
-        let deletion_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("")
-            .save(&conn)
-            .unwrap();
-        let deletion_node_id =
-            Node::create(&conn, &deletion_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let deletion = PathBlock {
-            node_id: deletion_node_id,
-            block_sequence: deletion_sequence.get_sequence(None, None).unwrap(),
-            sequence_start: 0,
-            sequence_end: 0,
-            path_start: 18,
-            path_end: 20,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(&conn, block_group_id, &path, 18, 20).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: deletion,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(&conn, &change).unwrap();
-
-        let all_sequences = BlockGroup::get_all_sequences(&conn, &block_group_id, false).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
     }
 
     #[test]
@@ -3384,1068 +2225,5 @@ mod tests {
                 },
             ],
         );
-    }
-
-    #[test]
-    fn test_changes_against_derivative_blockgroups() {
-        let conn = &get_connection(None).unwrap();
-        let (_block_group_id, _path) = setup_block_group(conn);
-        let _new_sample = Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "child",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        let new_bg_id = get_single_bg_id(conn, "test", "child", "chr1", vec!["test".to_string()]);
-        let new_path = Path::query(
-            conn,
-            "select * from paths where block_group_id = ?1",
-            rusqlite::params!(SQLValue::from(new_bg_id)),
-        );
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let region = ResolvedGenRegion::from_path(conn, new_bg_id, &new_path[0], 7, 15).unwrap();
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: false,
-        };
-
-        // note we are making our change against the new blockgroup, and not the parent blockgroup
-        BlockGroup::insert_change(conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(conn, &new_bg_id, true).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec!["AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),])
-        );
-
-        // Now, we make a change against another descendant
-        let _new_sample = Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "grandchild",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        let gc_bg_id = get_single_bg_id(
-            conn,
-            "test",
-            "grandchild",
-            "chr1",
-            vec!["child".to_string()],
-        );
-        let _new_path = Path::query(
-            conn,
-            "select * from paths where block_group_id = ?1",
-            rusqlite::params!(SQLValue::from(gc_bg_id)),
-        );
-
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 15,
-            strand: Strand::Forward,
-        };
-        let gc_bg = BlockGroup::get_by_id(conn, &gc_bg_id, None).unwrap();
-        let gc_region = ResolvedGenRegion {
-            block_group: gc_bg,
-            path: None,
-            accession: None,
-            annotation: None,
-            kind: ResolvedRegionKind::BlockGroup,
-            anchor_start: 0,
-            anchor_end: 0,
-            feature_length: 0,
-            start: 7,
-            end: 15,
-            start_anchors: None,
-            end_anchors: None,
-            remove_ambiguous_positions: true,
-        };
-        let change = BlockGroupChange {
-            region: gc_region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: false,
-        };
-        BlockGroup::insert_change(conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(conn, &gc_bg_id, true).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec!["AAAAAAANNNNTCCCCCCCCCCGGGGGGGGGG".to_string(),])
-        );
-    }
-
-    #[test]
-    fn test_changes_against_derivative_diploid_blockgroups() {
-        // This test ensures that if we have heterozygous changes that do not introduce frameshifts,
-        // we can modify regions downstream of them.
-        let conn = &get_connection(None).unwrap();
-        let (_block_group_id, _path) = setup_block_group(conn);
-        let _new_sample = Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "child",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        let new_bg_id = get_single_bg_id(conn, "test", "child", "chr1", vec!["test".to_string()]);
-        let _new_path = Path::query(
-            conn,
-            "select * from paths where block_group_id = ?1",
-            params![new_bg_id],
-        );
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 11,
-            strand: Strand::Forward,
-        };
-        let bg = BlockGroup::get_by_id(conn, &new_bg_id, None).unwrap();
-        let region = ResolvedGenRegion {
-            block_group: bg,
-            path: None,
-            accession: None,
-            annotation: None,
-            kind: ResolvedRegionKind::BlockGroup,
-            anchor_start: 0,
-            anchor_end: 0,
-            feature_length: 0,
-            start: 7,
-            end: 11,
-            start_anchors: None,
-            end_anchors: None,
-            remove_ambiguous_positions: true,
-        };
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(conn, &new_bg_id, true).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAANNNNTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-
-        // Now, we make a change against another descendant
-        let _new_sample = Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "grandchild",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        let gc_bg_id = get_single_bg_id(
-            conn,
-            "test",
-            "grandchild",
-            "chr1",
-            vec!["child".to_string()],
-        );
-        let _new_path = Path::query(
-            conn,
-            "select * from paths where block_group_id = ?1",
-            params![gc_bg_id],
-        );
-
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(conn)
-            .unwrap();
-        let insert_node_id = Node::create(
-            conn,
-            &insert_sequence.hash,
-            &HashId::convert_str("new-hash"),
-        )
-        .unwrap();
-
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 20,
-            path_end: 24,
-            strand: Strand::Forward,
-        };
-        let gc_bg = BlockGroup::get_by_id(conn, &gc_bg_id, None).unwrap();
-        let gc_region = ResolvedGenRegion {
-            block_group: gc_bg,
-            path: None,
-            accession: None,
-            annotation: None,
-            kind: ResolvedRegionKind::BlockGroup,
-            anchor_start: 0,
-            anchor_end: 0,
-            feature_length: 0,
-            start: 20,
-            end: 24,
-            start_anchors: None,
-            end_anchors: None,
-            remove_ambiguous_positions: true,
-        };
-        let change = BlockGroupChange {
-            region: gc_region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(conn, &gc_bg_id, true).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAANNNNTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTTNNNNCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAANNNNTTTTTTTTTNNNNCCCCCCGGGGGGGGGG".to_string()
-            ])
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_prohibits_out_of_frame_changes_against_derivative_diploid_blockgroups() {
-        // This test ensures that we do not allow ambiguous changes by coordinates
-        let conn = &get_connection(None).unwrap();
-        let (_block_group_id, _path) = setup_block_group(conn);
-        let _new_sample = Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "child",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        let new_bg_id = get_single_bg_id(conn, "test", "child", "chr1", vec!["test".to_string()]);
-        let _new_path = Path::query(
-            conn,
-            "select * from paths where block_group_id = ?1",
-            rusqlite::params!(SQLValue::from(new_bg_id)),
-        );
-        // This is a heterozygous replacement of 5 bases with 4 bases, so positions
-        // downstream of this are not addressable.
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(conn, &insert_sequence.hash, &HashId::convert_str("1")).unwrap();
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 7,
-            path_end: 12,
-            strand: Strand::Forward,
-        };
-        let bg = BlockGroup::get_by_id(conn, &new_bg_id, None).unwrap();
-        let region = ResolvedGenRegion {
-            block_group: bg,
-            path: None,
-            accession: None,
-            annotation: None,
-            kind: ResolvedRegionKind::BlockGroup,
-            anchor_start: 0,
-            anchor_end: 0,
-            feature_length: 0,
-            start: 7,
-            end: 12,
-            start_anchors: None,
-            end_anchors: None,
-            remove_ambiguous_positions: true,
-        };
-        let change = BlockGroupChange {
-            region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-
-        // note we are making our change against the new blockgroup, and not the parent blockgroup
-        BlockGroup::insert_change(conn, &change).unwrap();
-        let all_sequences = BlockGroup::get_all_sequences(conn, &new_bg_id, true).unwrap();
-        assert_eq!(
-            all_sequences,
-            HashSet::from_iter(vec![
-                "AAAAAAANNNNTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-            ])
-        );
-
-        // Now, we make a change against another descendant and get an error
-        let _new_sample = Sample::get_or_create(
-            conn,
-            NewSample {
-                name: "grandchild",
-                ..Default::default()
-            },
-        )
-        .unwrap();
-        let gc_bg_id = get_single_bg_id(
-            conn,
-            "test",
-            "grandchild",
-            "chr1",
-            vec!["child".to_string()],
-        );
-        let _new_path = Path::query(
-            conn,
-            "select * from paths where block_group_id = ?1",
-            rusqlite::params!(SQLValue::from(gc_bg_id)),
-        );
-
-        let insert_sequence = Sequence::new()
-            .sequence_type("DNA")
-            .sequence("NNNN")
-            .save(conn)
-            .unwrap();
-        let insert_node_id =
-            Node::create(conn, &insert_sequence.hash, &HashId::pad_str("new-hash")).unwrap();
-
-        let insert = PathBlock {
-            node_id: insert_node_id,
-            block_sequence: insert_sequence.get_sequence(0, 4).unwrap(),
-            sequence_start: 0,
-            sequence_end: 4,
-            path_start: 20,
-            path_end: 24,
-            strand: Strand::Forward,
-        };
-        let gc_bg = BlockGroup::get_by_id(conn, &gc_bg_id, None).unwrap();
-        let gc_region = ResolvedGenRegion {
-            block_group: gc_bg,
-            path: None,
-            accession: None,
-            annotation: None,
-            kind: ResolvedRegionKind::BlockGroup,
-            anchor_start: 0,
-            anchor_end: 0,
-            feature_length: 0,
-            start: 20,
-            end: 24,
-            start_anchors: None,
-            end_anchors: None,
-            remove_ambiguous_positions: true,
-        };
-        let change = BlockGroupChange {
-            region: gc_region,
-            path_accession: None,
-            block: insert,
-            chromosome_index: 1,
-            phased: 0,
-            preserve_edge: true,
-        };
-        BlockGroup::insert_change(conn, &change).unwrap();
-    }
-
-    mod test_derive_subgraph {
-
-        use super::*;
-        use crate::{
-            node::Node,
-            sequence::Sequence,
-            test_helpers::{get_connection, setup_block_group},
-        };
-
-        #[test]
-        fn test_derive_subgraph_one_insertion() {
-            /*
-            AAAAAAAAAA -> TTTTTTTTTT -> CCCCCCCCCC -> GGGGGGGGGG
-                              \-> AAAAAAAA ->/
-            Subgraph range:  |-----------------|
-            Sequences of the subgraph are TAAAAAAAAC, TTTTTCCCCC
-             */
-            let conn = &get_connection(None).unwrap();
-            let (block_group1_id, original_path) = setup_block_group(conn);
-
-            let intervaltree = original_path.intervaltree(conn).unwrap();
-            let insert_start_node_id = intervaltree.query_point(16).next().unwrap().value.node_id;
-            let insert_end_node_id = intervaltree.query_point(24).next().unwrap().value.node_id;
-
-            let insert_sequence = Sequence::new()
-                .sequence_type("DNA")
-                .sequence("AAAAAAAA")
-                .save(conn)
-                .unwrap();
-            let insert_node_id = Node::create(
-                conn,
-                &insert_sequence.hash,
-                &HashId(calculate_hash(&format!(
-                    "test-insert-a-node.{}",
-                    insert_sequence.hash
-                ))),
-            )
-            .unwrap();
-            let edge_into_insert = Edge::create(
-                conn,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-                insert_node_id,
-                0,
-                Strand::Forward,
-            )
-            .unwrap();
-            let edge_out_of_insert = Edge::create(
-                conn,
-                insert_node_id,
-                8,
-                Strand::Forward,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_1 = Edge::create(
-                conn,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_2 = Edge::create(
-                conn,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-
-            let edge_ids = [
-                &edge_into_insert.id,
-                &edge_out_of_insert.id,
-                &ref_heal_1.id,
-                &ref_heal_2.id,
-            ];
-            let block_group_edges = edge_ids
-                .iter()
-                .enumerate()
-                .map(|(i, edge_id)| BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: *(*edge_id),
-                    chromosome_index: if i < 2 { 1 } else { 0 },
-                    phased: 0,
-                })
-                .collect::<Vec<BlockGroupEdgeData>>();
-            BlockGroupEdge::bulk_create(conn, &block_group_edges);
-
-            let insert_path = original_path
-                .new_path_with(conn, 16, 24, &edge_into_insert, &edge_out_of_insert)
-                .unwrap();
-            assert_eq!(
-                insert_path.sequence(conn, None).unwrap(),
-                "AAAAAAAAAATTTTTTAAAAAAAACCCCCCGGGGGGGGGG"
-            );
-
-            let all_sequences =
-                BlockGroup::get_all_sequences(conn, &block_group1_id, false).unwrap();
-            assert_eq!(
-                all_sequences,
-                HashSet::from_iter(vec![
-                    "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTAAAAAAAACCCCCCGGGGGGGGGG".to_string(),
-                ])
-            );
-
-            let mut blocks = intervaltree
-                .query(Range { start: 15, end: 25 })
-                .map(|x| x.value)
-                .collect::<Vec<_>>();
-            blocks.sort_by_key(|a| a.start);
-            let start_block = blocks[0];
-            let start_node_coordinate = 15 - start_block.start + start_block.sequence_start;
-            let end_block = blocks[blocks.len() - 1];
-            let end_node_coordinate = 25 - end_block.start + end_block.sequence_start;
-
-            let block_group2 = create_bg(conn, "test", "test", "chr1.1");
-            let node_count_before = Node::query(conn, "SELECT * FROM nodes", params![]).len();
-            BlockGroup::derive_subgraph(
-                conn,
-                &block_group1_id,
-                &start_block,
-                &end_block,
-                start_node_coordinate,
-                end_node_coordinate,
-                &block_group2.id,
-                true,
-            )
-            .unwrap();
-            let node_count_after = Node::query(conn, "SELECT * FROM nodes", params![]).len();
-            assert_eq!(node_count_after, node_count_before);
-            let all_sequences2 =
-                BlockGroup::get_all_sequences(conn, &block_group2.id, false).unwrap();
-            assert_eq!(
-                all_sequences2,
-                HashSet::from_iter(vec!["TTTTTCCCCC".to_string(), "TAAAAAAAAC".to_string(),])
-            );
-        }
-
-        #[test]
-        fn test_derive_subgraph_two_independent_insertions() {
-            /*
-            AAAAAAAAAA -> TTTTTTTTTT -> CCCCCCCCCC -----> GGGGGGGGGG
-                                  \-> AAAAAAAA ->/  \->TTTTTTTT -/
-            Subgraph range:     |----------------------------------|
-             */
-            let conn = &get_connection(None).unwrap();
-            let (block_group1_id, original_path) = setup_block_group(conn);
-
-            let intervaltree = original_path.intervaltree(conn).unwrap();
-            let insert_start_node_id = intervaltree.query_point(16).next().unwrap().value.node_id;
-            let insert_end_node_id = intervaltree.query_point(24).next().unwrap().value.node_id;
-
-            let insert_sequence = Sequence::new()
-                .sequence_type("DNA")
-                .sequence("AAAAAAAA")
-                .save(conn)
-                .unwrap();
-            let insert_node_id = Node::create(
-                conn,
-                &insert_sequence.hash,
-                &HashId(calculate_hash(&format!(
-                    "test-insert-a-node.{}",
-                    insert_sequence.hash
-                ))),
-            )
-            .unwrap();
-            let edge_into_insert = Edge::create(
-                conn,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-                insert_node_id,
-                0,
-                Strand::Forward,
-            )
-            .unwrap();
-            let edge_out_of_insert = Edge::create(
-                conn,
-                insert_node_id,
-                8,
-                Strand::Forward,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_1 = Edge::create(
-                conn,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_2 = Edge::create(
-                conn,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-
-            let edge_ids = [
-                &edge_into_insert.id,
-                &edge_out_of_insert.id,
-                &ref_heal_1.id,
-                &ref_heal_2.id,
-            ];
-            let block_group_edges = edge_ids
-                .iter()
-                .enumerate()
-                .map(|(i, edge_id)| BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: *(*edge_id),
-                    chromosome_index: if i < 2 { 1 } else { 0 },
-                    phased: 0,
-                })
-                .collect::<Vec<BlockGroupEdgeData>>();
-            BlockGroupEdge::bulk_create(conn, &block_group_edges);
-
-            let insert_path = original_path
-                .new_path_with(conn, 16, 24, &edge_into_insert, &edge_out_of_insert)
-                .unwrap();
-            assert_eq!(
-                insert_path.sequence(conn, None).unwrap(),
-                "AAAAAAAAAATTTTTTAAAAAAAACCCCCCGGGGGGGGGG"
-            );
-
-            let insert2_start_node_id = intervaltree.query_point(28).next().unwrap().value.node_id;
-            let insert2_end_node_id = intervaltree.query_point(32).next().unwrap().value.node_id;
-
-            let insert2_sequence = Sequence::new()
-                .sequence_type("DNA")
-                .sequence("TTTTTTTT")
-                .save(conn)
-                .unwrap();
-            let insert2_node_id = Node::create(
-                conn,
-                &insert2_sequence.hash,
-                &HashId(calculate_hash(&format!(
-                    "test-insert-t-node.{}",
-                    insert2_sequence.hash
-                ))),
-            )
-            .unwrap();
-            let edge_into_insert2 = Edge::create(
-                conn,
-                insert2_start_node_id,
-                6,
-                Strand::Forward,
-                insert2_node_id,
-                0,
-                Strand::Forward,
-            )
-            .unwrap();
-            let edge_out_of_insert2 = Edge::create(
-                conn,
-                insert2_node_id,
-                8,
-                Strand::Forward,
-                insert2_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_1 = Edge::create(
-                conn,
-                insert2_start_node_id,
-                6,
-                Strand::Forward,
-                insert2_start_node_id,
-                6,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_2 = Edge::create(
-                conn,
-                insert2_end_node_id,
-                4,
-                Strand::Forward,
-                insert2_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-
-            let edge_ids = [
-                &edge_into_insert2.id,
-                &edge_out_of_insert2.id,
-                &ref_heal_1.id,
-                &ref_heal_2.id,
-            ];
-            let block_group_edges = edge_ids
-                .iter()
-                .enumerate()
-                .map(|(i, edge_id)| BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: *(*edge_id),
-                    chromosome_index: if i < 2 { 1 } else { 0 },
-                    phased: 0,
-                })
-                .collect::<Vec<BlockGroupEdgeData>>();
-            BlockGroupEdge::bulk_create(conn, &block_group_edges);
-
-            let insert2_path = insert_path
-                .new_path_with(conn, 28, 32, &edge_into_insert2, &edge_out_of_insert2)
-                .unwrap();
-            assert_eq!(
-                insert2_path.sequence(conn, None).unwrap(),
-                "AAAAAAAAAATTTTTTAAAAAAAACCTTTTTTTTGGGGGG"
-            );
-
-            let all_sequences =
-                BlockGroup::get_all_sequences(conn, &block_group1_id, false).unwrap();
-            assert_eq!(
-                all_sequences,
-                HashSet::from_iter(vec![
-                    "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTAAAAAAAACCCCCCGGGGGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTTTTTCCCCCCTTTTTTTTGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTAAAAAAAACCTTTTTTTTGGGGGG".to_string(),
-                ])
-            );
-
-            let mut blocks = intervaltree
-                .query(Range { start: 15, end: 36 })
-                .map(|x| x.value)
-                .collect::<Vec<_>>();
-            blocks.sort_by_key(|a| a.start);
-            let start_block = blocks[0];
-            let start_node_coordinate = 15 - start_block.start + start_block.sequence_start;
-            let end_block = blocks[blocks.len() - 1];
-            let end_node_coordinate = 36 - end_block.start + end_block.sequence_start;
-
-            let block_group2 = create_bg(conn, "test", "test", "chr1.1");
-            BlockGroup::derive_subgraph(
-                conn,
-                &block_group1_id,
-                &start_block,
-                &end_block,
-                start_node_coordinate,
-                end_node_coordinate,
-                &block_group2.id,
-                true,
-            )
-            .unwrap();
-            let all_sequences2 =
-                BlockGroup::get_all_sequences(conn, &block_group2.id, false).unwrap();
-            assert_eq!(
-                all_sequences2,
-                HashSet::from_iter(vec![
-                    "TTTTTCCCCCCCCCCGGGGGG".to_string(),
-                    "TAAAAAAAACCCCCCGGGGGG".to_string(),
-                    "TTTTTCCCCCCTTTTTTTTGG".to_string(),
-                    "TAAAAAAAACCTTTTTTTTGG".to_string(),
-                ])
-            );
-        }
-
-        #[test]
-        fn test_derive_subgraph_two_independent_insertions_and_one_deletion() {
-            /*
-                       /--------------------------------------------\  (<-- Deletion edge)
-            AAAAAAAAAA -> TTTTTTTTTT -> CCCCCCCCCC -----> GGGGGGGGGG
-                                  \-> AAAAAAAA ->/  \->TTTTTTTT -/
-            Subgraph range: |----------------------------------|
-
-            Confirms that deletion edge is ignored and not added to subgraph
-             */
-            let conn = &get_connection(None).unwrap();
-            let (block_group1_id, original_path) = setup_block_group(conn);
-
-            let intervaltree = original_path.intervaltree(conn).unwrap();
-            let insert_start_node_id = intervaltree.query_point(16).next().unwrap().value.node_id;
-            let insert_end_node_id = intervaltree.query_point(24).next().unwrap().value.node_id;
-
-            let insert_sequence = Sequence::new()
-                .sequence_type("DNA")
-                .sequence("AAAAAAAA")
-                .save(conn)
-                .unwrap();
-            let insert_node_id = Node::create(
-                conn,
-                &insert_sequence.hash,
-                &HashId(calculate_hash(&format!(
-                    "test-insert-a-node.{}",
-                    insert_sequence.hash
-                ))),
-            )
-            .unwrap();
-            let edge_into_insert = Edge::create(
-                conn,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-                insert_node_id,
-                0,
-                Strand::Forward,
-            )
-            .unwrap();
-            let edge_out_of_insert = Edge::create(
-                conn,
-                insert_node_id,
-                8,
-                Strand::Forward,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_1 = Edge::create(
-                conn,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-                insert_start_node_id,
-                6,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_2 = Edge::create(
-                conn,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-                insert_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-
-            let edge_ids = [
-                &edge_into_insert.id,
-                &edge_out_of_insert.id,
-                &ref_heal_1.id,
-                &ref_heal_2.id,
-            ];
-            let block_group_edges = edge_ids
-                .iter()
-                .map(|edge_id| BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: *(*edge_id),
-                    chromosome_index: NO_CHROMOSOME_INDEX,
-                    phased: 0,
-                })
-                .collect::<Vec<BlockGroupEdgeData>>();
-            BlockGroupEdge::bulk_create(conn, &block_group_edges);
-
-            let insert_path = original_path
-                .new_path_with(conn, 16, 24, &edge_into_insert, &edge_out_of_insert)
-                .unwrap();
-            assert_eq!(
-                insert_path.sequence(conn, None).unwrap(),
-                "AAAAAAAAAATTTTTTAAAAAAAACCCCCCGGGGGGGGGG"
-            );
-
-            let insert2_start_node_id = intervaltree.query_point(28).next().unwrap().value.node_id;
-            let insert2_end_node_id = intervaltree.query_point(32).next().unwrap().value.node_id;
-
-            let insert2_sequence = Sequence::new()
-                .sequence_type("DNA")
-                .sequence("TTTTTTTT")
-                .save(conn)
-                .unwrap();
-            let insert2_node_id = Node::create(
-                conn,
-                &insert2_sequence.hash,
-                &HashId(calculate_hash(&format!(
-                    "test-insert-t-node.{}",
-                    insert2_sequence.hash
-                ))),
-            )
-            .unwrap();
-            let edge_into_insert2 = Edge::create(
-                conn,
-                insert2_start_node_id,
-                6,
-                Strand::Forward,
-                insert2_node_id,
-                0,
-                Strand::Forward,
-            )
-            .unwrap();
-            let edge_out_of_insert2 = Edge::create(
-                conn,
-                insert2_node_id,
-                8,
-                Strand::Forward,
-                insert2_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_1 = Edge::create(
-                conn,
-                insert2_start_node_id,
-                6,
-                Strand::Forward,
-                insert2_start_node_id,
-                6,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_2 = Edge::create(
-                conn,
-                insert2_end_node_id,
-                4,
-                Strand::Forward,
-                insert2_end_node_id,
-                4,
-                Strand::Forward,
-            )
-            .unwrap();
-
-            let edge_ids = [
-                &edge_into_insert2.id,
-                &edge_out_of_insert2.id,
-                &ref_heal_1.id,
-                &ref_heal_2.id,
-            ];
-            let block_group_edges = edge_ids
-                .iter()
-                .map(|edge_id| BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: *(*edge_id),
-                    chromosome_index: NO_CHROMOSOME_INDEX,
-                    phased: 0,
-                })
-                .collect::<Vec<BlockGroupEdgeData>>();
-            BlockGroupEdge::bulk_create(conn, &block_group_edges);
-
-            let insert2_path = insert_path
-                .new_path_with(conn, 28, 32, &edge_into_insert2, &edge_out_of_insert2)
-                .unwrap();
-            assert_eq!(
-                insert2_path.sequence(conn, None).unwrap(),
-                "AAAAAAAAAATTTTTTAAAAAAAACCTTTTTTTTGGGGGG"
-            );
-
-            let deletion_end_node_id = intervaltree.query_point(38).next().unwrap().value.node_id;
-            let deletion_edge = Edge::create(
-                conn,
-                insert_node_id,
-                8,
-                Strand::Forward,
-                deletion_end_node_id,
-                8,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_1 = Edge::create(
-                conn,
-                insert_node_id,
-                8,
-                Strand::Forward,
-                insert_node_id,
-                8,
-                Strand::Forward,
-            )
-            .unwrap();
-            let ref_heal_2 = Edge::create(
-                conn,
-                deletion_end_node_id,
-                8,
-                Strand::Forward,
-                deletion_end_node_id,
-                8,
-                Strand::Forward,
-            )
-            .unwrap();
-            let block_group_edges = [
-                BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: deletion_edge.id,
-                    chromosome_index: NO_CHROMOSOME_INDEX,
-                    phased: 0,
-                },
-                BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: ref_heal_1.id,
-                    chromosome_index: NO_CHROMOSOME_INDEX,
-                    phased: 0,
-                },
-                BlockGroupEdgeData {
-                    block_group_id: block_group1_id,
-                    edge_id: ref_heal_2.id,
-                    chromosome_index: NO_CHROMOSOME_INDEX,
-                    phased: 0,
-                },
-            ];
-            BlockGroupEdge::bulk_create(conn, &block_group_edges);
-
-            let all_sequences =
-                BlockGroup::get_all_sequences(conn, &block_group1_id, false).unwrap();
-            assert_eq!(
-                all_sequences,
-                HashSet::from_iter(vec![
-                    "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTAAAAAAAACCCCCCGGGGGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTTTTTCCCCCCTTTTTTTTGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTAAAAAAAACCTTTTTTTTGGGGGG".to_string(),
-                    "AAAAAAAAAATTTTTTAAAAAAAAGG".to_string(), // Sequence including deletion
-                ])
-            );
-
-            let mut blocks = intervaltree
-                .query(Range { start: 15, end: 36 })
-                .map(|x| x.value)
-                .collect::<Vec<_>>();
-            blocks.sort_by_key(|a| a.start);
-            let start_block = blocks[0];
-            let start_node_coordinate = 15 - start_block.start + start_block.sequence_start;
-            let end_block = blocks[blocks.len() - 1];
-            let end_node_coordinate = 36 - end_block.start + end_block.sequence_start;
-
-            let block_group2 = create_bg(conn, "test", "test", "chr1.1");
-            BlockGroup::derive_subgraph(
-                conn,
-                &block_group1_id,
-                &start_block,
-                &end_block,
-                start_node_coordinate,
-                end_node_coordinate,
-                &block_group2.id,
-                true,
-            )
-            .unwrap();
-            let all_sequences2 =
-                BlockGroup::get_all_sequences(conn, &block_group2.id, false).unwrap();
-            assert_eq!(
-                all_sequences2,
-                // The deletion is not included in the cloned subgraph since one end of it is
-                // outside the specified range
-                HashSet::from_iter(vec![
-                    "TTTTTCCCCCCCCCCGGGGGG".to_string(),
-                    "TAAAAAAAACCCCCCGGGGGG".to_string(),
-                    "TTTTTCCCCCCTTTTTTTTGG".to_string(),
-                    "TAAAAAAAACCTTTTTTTTGG".to_string(),
-                ])
-            );
-        }
     }
 }
